Geen idee meer, iidg veel

+++ src/main/java/nl/saxion/game/Bloodspire/Classes/Enemy.java
@@ -2,10 +2,10 @@ package nl.saxion.game.Bloodspire.Classes;
 
 public class Enemy {
     String enemyName = "";
-    double hitPoints = 50;
-    double attackSpeed = 1;
-    double attackDamage = 1;
-    double defense = 0;
-    double experiencePoints = 1;
+    int hitPoints = 50;
+    int attackSpeed = 1;
+    int attackDamage = 1;
+    int defense = 0;
+    int experiencePoints = 1;
     int enemyID = 0;
 }
